docker-mnist-service/requirements.txt
tensorflow==1.13.1
polyaxon-client==0.4.4
imageio==2.5.0
pillow==6.0.0
flask==1.0.2
